fix UI

+++ lib/Model/Data/Board/reply.dart
@@ -1,12 +1,13 @@
 import 'dart:convert';
+import 'dart:ffi';
 
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 class replyModel {
-  int? _replyId;
+  Long? _replyId;
   String? _replyContents;
   String? _replyWriter;
 
-  replyModel({int? replyId, String? replyContents, String? replyWriter})
+  replyModel({Long? replyId, String? replyContents, String? replyWriter})
       : _replyId = replyId,
         _replyContents = replyContents,
         _replyWriter = replyWriter;
@@ -20,7 +21,7 @@ class replyModel {
 
   factory replyModel.fromMap(Map<String, dynamic> map) {
     return replyModel(
-      replyId: map['replyId'] as int,
+      replyId: map['replyId'] as Long,
       replyContents: map['contents'] as String,
       replyWriter: map['writer'] as String,
     );
